Re-re-re-factor. analyze.py is back, YANK now has its own regiment of Analyzers which are subclasses of the Multistate Analyzers
Yank-specific functions have been taken out of the multistate module, which is now even more independent.

The ease of extending the analyzers is also proof-of-concept for the independence of the multistate module
