+++ Yank/multistate/analyzers/analyzerutils.py
@@ -17,13 +17,10 @@ Utilities and common functions for MultiStateSampler analyzers.
 # MODULE IMPORTS
 # =============================================================================================
 
-import re
 import logging
 import numpy as np
 
 from pymbar import timeseries  # for statistical inefficiency analysis
-
-from typing import Union, Optional
 
 logger = logging.getLogger(__name__)
 
@@ -34,7 +31,6 @@ __all__ = [
     'get_equilibration_data_per_sample',
     'remove_unequilibrated_data',
     'subsample_data_along_axis',
-    'ObservablesRegistry'
 ]
 
 
@@ -178,239 +174,3 @@ def subsample_data_along_axis(data, subsample_rate, axis):
     indices = timeseries.subsampleCorrelatedData(np.zeros(data_shape[axis]), g=subsample_rate)
     subsampled_data = np.take(cast_data, indices, axis=axis)
     return subsampled_data
-
-
-# =============================================================================================
-# MODULE CLASSES
-# =============================================================================================
-
-class ObservablesRegistry(object):
-    """
-    Registry of computable observables.
-
-    This is a hidden class accessed by the :class:`PhaseAnalyzer` objects to check
-    which observables can be computed, and then provide a regular categorization of them. This is a static registry.
-
-    To define your own methods:
-    1) Choose a unique observable name.
-    2) Categorize the observable in one of the following ways by adding to the list in the "observables_X" method:
-     2a) "defined_by_phase": Depends on the Phase as a whole (state independent)
-     2b) "defined_by_single_state": Computed entirely from one state, e.g. Radius of Gyration
-     2c) "defined_by_two_states": Property is relative to some reference state, such as Free Energy Difference
-    3) Optionally categorize the error category calculation in the "observables_with_error_adding_Y" methods
-       If not placed in an error category, the observable will be assumed not to carry error
-       Examples: A, B, C are the observable in 3 phases, eA, eB, eC are the error of the observable in each phase
-     3a) "linear": Error between phases adds linearly.
-        If C = A + B, eC = eA + eB
-     3b) "quadrature": Error between phases adds in the square.
-        If C = A + B, eC = sqrt(eA^2 + eB^2)
-    4) Finally, to add this observable to the phase, implement a "get_{method name}" method to the subclass of
-       :class:`YankPhaseAnalyzer`. Any :class:`MultiPhaseAnalyzer` composed of this phase will automatically have the
-       "get_{method name}" if all other phases in the :class:`MultiPhaseAnalyzer` have the same method.
-    """
-
-    def __init__(self):
-        """Register Defaults"""
-        # Create empty registry
-        self._observables = {'two_state': set(),
-                             'one_state': set(),
-                             'phase': set()}
-        self._errors = {'quad': set(),
-                        'linear': set(),
-                        None: set()}
-
-    def register_two_state_observable(self, name: str,
-                                      error_class: Optional[str]=None,
-                                      re_register: bool=False):
-        """
-        Register a new two state observable, or re-register an existing one.
-
-        Parameters
-        ----------
-        name: str
-            Name of the observable, will be cast to all lower case and spaces replaced with underscores
-        error_class: "quad", "linear", or None
-            How the error of the observable is computed when added with other errors from the same observable.
-
-            * "quad": Adds in the quadrature, Observable C = A + B, Error eC = sqrt(eA**2 + eB**2)
-
-            * "linear": Adds linearly,  Observable C = A + B, Error eC = eA + eB
-
-            * None: Does not carry error
-
-        re_register: bool, optional, Default: False
-            Re-register an existing observable
-        """
-
-        self._register_observable(name, "two_state", error_class, re_register=re_register)
-
-    def register_one_state_observable(self, name: str,
-                                      error_class: Optional[str]=None,
-                                      re_register: bool=False):
-        """
-        Register a new one state observable, or re-register an existing one.
-
-        Parameters
-        ----------
-        name: str
-            Name of the observable, will be cast to all lower case and spaces replaced with underscores
-        error_class: "quad", "linear", or None
-            How the error of the observable is computed when added with other errors from the same observable.
-
-            * "quad": Adds in the quadrature, Observable C = A + B, Error eC = sqrt(eA**2 + eB**2)
-
-            * "linear": Adds linearly,  Observable C = A + B, Error eC = eA + eB
-
-            * None: Does not carry error
-
-        re_register: bool, optional, Default: False
-            Re-register an existing observable
-        """
-
-        self._register_observable(name, "one_state", error_class, re_register=re_register)
-
-    def register_phase_observable(self, name: str,
-                                  error_class: Optional[str]=None,
-                                  re_register: bool=False):
-        """
-        Register a new observable defined by phaee, or re-register an existing one.
-
-        Parameters
-        ----------
-        name: str
-            Name of the observable, will be cast to all lower case and spaces replaced with underscores
-        error_class: "quad", "linear", or None
-            How the error of the observable is computed when added with other errors from the same observable.
-
-            * "quad": Adds in the quadrature, Observable C = A + B, Error eC = sqrt(eA**2 + eB**2)
-
-            * "linear": Adds linearly,  Observable C = A + B, Error eC = eA + eB
-
-            * None: Does not carry error
-
-        re_register: bool, optional, Default: False
-            Re-register an existing observable
-        """
-
-        self._register_observable(name, "phase", error_class, re_register=re_register)
-
-    ########################
-    # Define the observables
-    ########################
-    @property
-    def observables(self) -> tuple:
-        """
-        Set of observables which are derived from the subsets below
-        """
-        observables = set()
-        for subset_key in self._observables:
-            observables |= self._observables[subset_key]
-        return tuple(observables)
-
-    # ------------------------------------------------
-    # Exclusive Observable categories
-    # The intersection of these should be the null set
-    # ------------------------------------------------
-
-    @property
-    def observables_defined_by_two_states(self) -> tuple:
-        """
-        Observables that require an i and a j state to define the observable accurately between phases
-        """
-        return self._get_observables('two_state')
-
-    @property
-    def observables_defined_by_single_state(self) -> tuple:
-        """
-        Defined observables which are fully defined by a single state, and not by multiple states such as differences
-        """
-        return self._get_observables('one_state')
-
-    @property
-    def observables_defined_by_phase(self) -> tuple:
-        """
-        Observables which are defined by the phase as a whole, and not defined by any 1 or more states
-        e.g. Standard State Correction
-        """
-        return self._get_observables('phase')
-
-    ##########################################
-    # Define the observables which carry error
-    # This should be a subset of observables
-    ##########################################
-
-    @property
-    def observables_with_error(self) -> tuple:
-        """Determine which observables have error by inspecting the the error subsets"""
-        observables = set()
-        for subset_key in self._errors:
-            if subset_key is not None:
-                observables |= self._errors[subset_key]
-        return tuple(observables)
-
-    # ------------------------------------------------
-    # Exclusive Error categories
-    # The intersection of these should be the null set
-    # ------------------------------------------------
-
-    @property
-    def observables_with_error_adding_quadrature(self) -> tuple:
-        """Observable C = A + B, Error eC = sqrt(eA**2 + eB**2)"""
-        return self._get_errors('quad')
-
-    @property
-    def observables_with_error_adding_linear(self) -> tuple:
-        """Observable C = A + B, Error eC = eA + eB"""
-        return self._get_errors('linear')
-
-    @property
-    def observables_without_error(self) -> tuple:
-        return self._get_errors(None)
-
-    # ------------------
-    # Internal functions
-    # ------------------
-
-    def _get_observables(self, key) -> tuple:
-        return tuple(self._observables[key])
-
-    def _get_errors(self, key) -> tuple:
-        return tuple(self._errors[key])
-
-    @staticmethod
-    def _cast_observable_name(name) -> str:
-        return re.sub(" +", "_", name.lower())
-
-    def _register_observable(self, obs_name: str,
-                             obs_calc_class: str,
-                             obs_error_class: Union[None, str],
-                             re_register: bool=False):
-        obs_name = self._cast_observable_name(obs_name)
-        if not re_register and obs_name in self.observables:
-            raise ValueError("{} is already a registered observable! "
-                             "Consider setting re_register key!".format(obs_name))
-        self._check_obs_class(obs_calc_class)
-        self._check_obs_error_class(obs_error_class)
-        obs_name_set = {obs_name}  # set(single_object) throws an error, set(string) splits each char
-        # Throw out existing observable if present (set difference)
-        for obs_key in self._observables:
-            self._observables[obs_key] -= obs_name_set
-        for obs_err_key in self._errors:
-            self._errors[obs_err_key] -= obs_name_set
-        # Add new observable to correct classifiers (set union)
-        self._observables[obs_calc_class] |= obs_name_set
-        self._errors[obs_error_class] |= obs_name_set
-
-    def _check_obs_class(self, obs_class):
-        assert obs_class in self._observables, "{} not a known observable class!".format(obs_class)
-
-    def _check_obs_error_class(self, obs_error):
-        assert obs_error is None or obs_error in self._errors, \
-            "{} not a known observable error class!".format(obs_error)
-
-# Create a default registry and register some stock values
-ObservablesRegistry = ObservablesRegistry()
-ObservablesRegistry.register_two_state_observable('free_energy', error_class='quad')
-ObservablesRegistry.register_two_state_observable('entropy', error_class='quad')
-ObservablesRegistry.register_two_state_observable('enthalpy', error_class='quad')
-ObservablesRegistry.register_phase_observable('standard_state_correction')
