+++ Yank/tests/test_analyze.py
@@ -9,8 +9,9 @@ Test multistate.analyzers facility.
 # GLOBAL IMPORTS
 # =============================================================================================
 
-import copy
 import os
+import re
+import copy
 import shutil
 import tempfile
 
@@ -23,7 +24,7 @@ from simtk import unit
 
 from yank.yank import Topography
 from yank.multistate import MultiStateReporter, ReplicaExchangeSampler
-import yank.multistate.analyzers as analyze
+import yank.analyze as analyze
 
 # ==============================================================================
 # MODULE CONSTANTS
@@ -36,7 +37,7 @@ kB = unit.BOLTZMANN_CONSTANT_kB * unit.AVOGADRO_CONSTANT_NA  # Boltzmann constan
 # TEMPLATE PHASE CLASS WITH NO OBSERVABLES
 # ==============================================================================
 
-class BlankPhase(analyze.PhaseAnalyzer):
+class BlankPhase(analyze.YankPhaseAnalyzer):
     """Create a blank phase class with no get_X (observable) methods for testing the MultiPhaseAnalyzer"""
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
@@ -180,13 +181,13 @@ class TestPhaseAnalyzer(object):
 
     def test_repex_phase_initialize(self):
         """Test that the MultiState Phase analyzer initializes correctly"""
-        phase = analyze.ReplicaExchangeAnalyzer(self.reporter, name=self.repex_name)
+        phase = analyze.YankReplicaExchangeAnalyzer(self.reporter, name=self.repex_name)
         assert phase.reporter is self.reporter
         assert phase.name == self.repex_name
 
     def test_repex_mixing_stats(self):
         """Test that the MultiState Phase yields mixing stats that make sense"""
-        phase = analyze.ReplicaExchangeAnalyzer(self.reporter, name=self.repex_name)
+        phase = analyze.YankReplicaExchangeAnalyzer(self.reporter, name=self.repex_name)
         t, mu, g = phase.generate_mixing_statistics()
         # Output is the correct number of states
         assert t.shape == (self.n_states, self.n_states)
@@ -204,7 +205,7 @@ class TestPhaseAnalyzer(object):
 
         We do this in one function since the test for each part would be a bunch of repeated recreation of the phase
         """
-        phase = analyze.ReplicaExchangeAnalyzer(self.reporter, name=self.repex_name)
+        phase = analyze.YankReplicaExchangeAnalyzer(self.reporter, name=self.repex_name)
         u_sampled, u_unsampled = phase.get_states_energies()
         # Test energy output matches appropriate MBAR shapes
         assert u_sampled.shape == (self.n_states, self.n_states, self.n_steps)
@@ -215,7 +216,7 @@ class TestPhaseAnalyzer(object):
         discard = 1
         # Generate mbar semi-manually, use phases's static methods
         n_eq, g_t, Neff_max = pymbar.timeseries.detectEquilibration(u_n[discard:])
-        u_sampled_sub = analyze.remove_unequilibrated_data(u_sampled, n_eq, -1)
+        u_sampled_sub = analyze.analysis.remove_unequilibrated_data(u_sampled, n_eq, -1)
         # Make sure output from subsample is what we expect
         assert u_sampled_sub.shape == (self.n_states, self.n_states, Neff_max)
         # Generate MBAR from phase
@@ -245,7 +246,7 @@ class TestPhaseAnalyzer(object):
     def test_multi_phase(self):
         """Test MultiPhaseAnalysis"""
         # Create the phases
-        full_phase = analyze.ReplicaExchangeAnalyzer(self.reporter, name=self.repex_name)
+        full_phase = analyze.YankReplicaExchangeAnalyzer(self.reporter, name=self.repex_name)
         blank_phase = BlankPhase(self.reporter, name="blank")
         fe_phase = FreeEnergyPhase(self.reporter, name="fe")
         fes_phase = FEStandardStatePhase(self.reporter, name="fes")
@@ -309,3 +310,9 @@ class TestPhaseAnalyzer(object):
         solute_trajectory = analyze.extract_trajectory(self.reporter.filepath, state_index=0, keep_solvent=False)
         assert len(solute_trajectory) == self.n_steps - 1
         assert solute_trajectory.n_atoms == len(self.analysis_atoms)
+
+    def test_yank_registry(self):
+        """Test that observable registry is implemented correctly for a given class"""
+        phase = analyze.YankReplicaExchangeAnalyzer(self.reporter, name=self.repex_name)
+        observables = set(analyze.yank_registry.observables)
+        assert set(phase.observables) == observables
