+++ docs/api/multistate_api/analyzers_api.rst
@@ -3,3 +3,7 @@
 .. automodule:: yank.multistate.analyzers
    :members:
    :inherited-members:
+
+.. automodule:: yank.analyze
+   :members:
+   :inherited-members:
